Add files via upload
This project is simple to create. I'm a beginner in reactJS . So I decided to start the easy way. We already see the GOOGLE is the best example to understand the to-do List.  The to-do list items are rendered as a list (<ul>), with each item represented by a list item (<li>). The list item includes a button for marking the task as complete, the task text, and a delete button.